+++ docs/notes/mainWindowSrc-notes.js
