+++ src/utils/misc.js
@@ -1,14 +1,15 @@
-// utils.js
+// utils/misc.js
 // LayerRenamer
 //
 // created by Leonard Pauli, jun 2018
 // copyright © Leonard Pauli 2017-2018
 
-/* globals NSStringFromClass */
-
 import utils from 'sketch-utils'
 import {log as _logd} from 'string-from-object'
-import config from './config'
+import config from '../config'
+
+
+// logging
 
 const logd = (...args)=> config.app.isProduction || _logd(...args)
 const logns = ns=> {
@@ -27,30 +28,46 @@ const logns = ns=> {
 	}))
 }
 
-const getLayerKind = obj=> {
-	const objClass = obj.class()
-	if (objClass == MSLayerGroup)
-		return 'Group'
 
-	let str = NSStringFromClass(objClass)
-	if (str.substr(0, 2)=='MS') str = str.substr(2)
-	if (str.substr(-5)=='Group' || str.substr(-5)=='Layer')
-		str = str.substr(0, str.length-5)
+// strings
 
-	if (str=='SymbolInstance') return 'Symbol'
-	if (str.length>5 && (str.substr(-5)=='Shape' || str=='ShapePath')) return 'Path'
-
-	return str
-}
+const regexEscape = s=> s.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&')
 
 const paddStringToLength = (str, len, append = false, char = ' ')=> {
 	let txt = ''; for (let i = Math.max(0, len-str.length); i>0; i--) txt += char
 	return append? str+txt: txt+str
 }
 
+// transformStringCaseUsingFlags
+const transformCaseRE = /(?:\\([LUCKESTW])((?:.|\s)+?)(?=(?:\\[LUCKESTW])|$))/g
+const toKebabCase = s=> s
+	.replace(/([a-z])([A-Z])/g, (_, f, l)=> f+'-'+l) // convert camel/title case
+	.toLowerCase() // discard case information
+	.replace(/_(\w)/ig, (_, l)=> '-'+l) // convert snake case
+	.replace(/ (\w)/ig, (_, l)=> '-'+l) // convert word case
+const transformers = {
+	L: s=> s.toLowerCase(), // lowercase
+	U: s=> s.toUpperCase(), // UPPERCASE
+	C: s=> toKebabCase(s).replace(/-(\w)/ig, (_, l)=> l.toUpperCase()), // camelCase
+	K: s=> toKebabCase(s), // kebab-case
+	E: s=> s, // end
+	S: s=> toKebabCase(s).replace(/-(\w)/ig, (_, l)=> '_'+l), // snake_case
+	T: s=> toKebabCase(s).replace(/(^|-)(\w)/ig, (_1, _2, l)=> l.toUpperCase()), // TitleCase
+	W: s=> toKebabCase(s).replace(/-(\w)/ig, (_, l)=> ' '+l), // word case
+}
+// var str = "\\Klowercase \\KUPPERCASE \\KcamelCase \\Kkebab-case \\Kend \\Ksnake_case \\KTitleCase"
+// str += "\n\\LlowerCase \\UupperCase \\CcamelCase \\KkebabCase \\Eend \\SsnakeCase \\TtitleCase"
+const transformStringCaseUsingFlags = str=> str.replace(transformCaseRE, (all, flag, str)=> transformers[flag](str))
+
+
+// system
+
 const scriptDirGet = coscript=> coscript.env().scriptURL.path().stringByDeletingLastPathComponent().stringByDeletingLastPathComponent()+'/'
 
 const openUrl = url=> NSWorkspace.sharedWorkspace().openURL(NSURL.URLWithString(url))
+
+
+// storage
 
 const storageStringGet = ({namespace})=> ({
 	store: NSUserDefaults.alloc().initWithSuiteName(`plugin.sketch.${namespace}`),
@@ -73,14 +90,18 @@ const storageJSONGet = ({namespace})=> ({
 	set (k, v) { return this.store.set(k, JSON.stringify(v)) },
 })
 
-const regexEscape = s=> s.replace(/[-/\\^$*+?.()|[\]{}]/g, '\\$&')
+
+// export
 
 export {
 	logd, logns,
-	scriptDirGet,
-	getLayerKind,
-	paddStringToLength,
-	openUrl,
-	storageStringGet, storageJSONGet,
+
 	regexEscape,
+	paddStringToLength,
+	transformStringCaseUsingFlags,
+
+	openUrl,
+	scriptDirGet,
+	
+	storageStringGet, storageJSONGet,
 }
