+++ src/utils/showCursorPopupBrowserWindow.js
